changed isizes

+++ src/vertex_buffer.rs
@@ -130,9 +130,9 @@ impl<'a> GLByteBuffer<'a> {
         gl::CopyBufferSubData(
           gl::ARRAY_BUFFER,
           gl::ARRAY_BUFFER,
-          self.length as i64,
-          i as i64,
-          count as i64,
+          self.length,
+          i,
+          count,
         );
       }
     }
@@ -154,8 +154,8 @@ impl<'a> GLByteBuffer<'a> {
 
     gl::BufferSubData(
       gl::ARRAY_BUFFER,
-      idx as i64,
-      count as i64,
+      idx,
+      count,
       mem::transmute(vs)
     );
   }
